requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
docker==3.7.2
docker-pycreds==0.4.0
idna==2.8
idna-ssl==1.1.0
mohawk==1.0.0
multidict==4.5.2
pyyaml==5.1
requests==2.21.0
six==1.12.0
slugid==2.0.0
taskcluster==7.0.1
taskcluster-urls==11.0.0
typing-extensions==3.7.2
urllib3==1.24.1
websocket-client==0.56.0
yarl==1.3.0
